Add files via upload
A simple Player vs. Monster project that asks the user for a few inputs. Attacking the monster, restoring health, and a few additional options are available to the player. Includes JavaScript as its key focus. Functions, Arrow functions, and the linking of data from several JavaScript files are all implemented.